+++ src/termvisage/tui/__init__.py
@@ -9,7 +9,8 @@ from pathlib import Path
 from typing import Any, Dict, Iterable, Iterator, Tuple, Union
 
 import urwid
-from term_image.utils import lock_tty, write_tty
+from term_image.image import GraphicsImage
+from term_image.utils import get_cell_size, lock_tty, write_tty
 from term_image.widget import UrwidImageScreen
 
 from .. import logging, notify
@@ -29,6 +30,8 @@ def init(
     ImageClass: type,
 ) -> None:
     """Initializes the TUI"""
+    from . import keys
+
     global active, initialized
 
     if args.debug:
@@ -41,6 +44,8 @@ def init(
     main.NO_ANIMATION = args.no_anim
     main.RECURSIVE = args.recursive
     main.SHOW_HIDDEN = args.all
+    main.THUMBNAIL = args.thumbnail
+    main.THUMBNAIL_SIZE_PRODUCT = config_options.thumbnail_size**2
     main.ImageClass = ImageClass
     main.loop = Loop(
         main_widget, palette, UrwidImageScreen(), unhandled_input=process_input
@@ -52,6 +57,7 @@ def init(
     )
     render.FRAME_DURATION = args.frame_duration
     render.REPEAT = args.repeat
+    render.THUMBNAIL_CACHE_SIZE = config_options.thumbnail_cache
 
     images.sort(
         key=lambda x: sort_key_lexi(
@@ -67,6 +73,11 @@ def init(
             specs = getattr(render, f"{name}_style_specs")
             specs[ImageClass.style] += f"c{style_args['compress']}"
 
+    if issubclass(ImageClass, GraphicsImage):
+        # `get_cell_size()` may sometimes return `None` on terminals that don't
+        # implement the `TIOCSWINSZ` ioctl command. Hence, the `or (1, 2)`.
+        keys._prev_cell_size = get_cell_size() or (1, 2)
+
     Image._ti_alpha = (
         "#"
         if args.no_alpha
@@ -77,6 +88,8 @@ def init(
         )
     )
     Image._ti_grid_style_spec = render.grid_style_specs.get(ImageClass.style, "")
+    if main.THUMBNAIL:
+        Image._ti_update_grid_thumbnailing_threshold(keys._prev_cell_size)
 
     # daemon, to avoid having to check if the main process has been interrupted
     menu_scanner = logging.Thread(target=scan_dir_menu, name="MenuScanner", daemon=True)
@@ -87,6 +100,13 @@ def init(
         name="GridRenderManager",
         daemon=True,
     )
+    if main.THUMBNAIL:
+        grid_thumbnail_manager = logging.Thread(
+            target=render.manage_grid_thumbnails,
+            args=(config_options.thumbnail_size,),
+            name="GridThumbnailManager",
+            daemon=True,
+        )
     image_render_manager = logging.Thread(
         target=render.manage_image_renders,
         name="ImageRenderManager",
@@ -121,6 +141,8 @@ def init(
 
     menu_scanner.start()
     grid_scanner.start()
+    if main.THUMBNAIL:
+        grid_thumbnail_manager.start()
     grid_render_manager.start()
     image_render_manager.start()
     anim_render_manager.start()
@@ -129,6 +151,8 @@ def init(
         write_tty(f"{CSI}?1049h".encode())  # Switch to the alternate buffer
         next(main.displayer)
         main.loop.run()
+        if main.THUMBNAIL:
+            grid_thumbnail_manager.join()
         grid_render_manager.join()
         render.image_render_queue.put((None,) * 3)
         image_render_manager.join()
@@ -181,4 +205,5 @@ palette = [
     ("error", "", "", "", "bold", "#ff0000"),
     ("warning", "", "", "", "#ff0000,bold", ""),
     ("notif context", "", "", "", "#0000ff,bold", ""),
+    ("high-res", "", "", "", "#a07f00", ""),
 ]